# optimize readUntilPrompt