--- a/pkg/netrasp/reader.go
+++ b/pkg/netrasp/reader.go
@@ -1,6 +1,7 @@
 package netrasp
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -10,6 +11,10 @@ import (
 )
 
 var errRead = errors.New("reader error")
+var repl = strings.NewReplacer("\r\n", "\n", "\r", "\n")
+
+var minReadSize = 10000
+var maxReadSize = 10 * 1000 * 1000
 
 type contextReader struct {
 	ctx context.Context
@@ -55,7 +60,7 @@ func newContextReader(ctx context.Context, r io.Reader) io.Reader {
 }
 
 // readUntilPrompt reads until the specified prompt is found and returns the read data.
-func readUntilPrompt(ctx context.Context, r io.Reader, prompt *regexp.Regexp) (string, error) {
+func readUntilPromptOriginal(ctx context.Context, r io.Reader, prompt *regexp.Regexp) (string, error) {
 	var output string
 	r = newContextReader(ctx, r)
 	for {
@@ -80,4 +85,55 @@ func readUntilPrompt(ctx context.Context, r io.Reader, prompt *regexp.Regexp) (s
 	}
 
 	return output, nil
+}
+
+// readUntilPrompt reads until the specified prompt is found and returns the read data.
+func readUntilPromptWithStringsBuilder(ctx context.Context, r io.Reader, prompt *regexp.Regexp) (string, error) {
+	var output = strings.Builder{}
+	rc := newContextReader(ctx, r)
+	readSize := minReadSize
+	for {
+		buffer := make([]byte, readSize)
+
+		n, err := rc.Read(buffer)
+		if err != nil {
+			return "", fmt.Errorf("error reading output from device %w: %v", errRead, err)
+		}
+		if readSize == n && readSize < maxReadSize {
+			readSize *= 2
+		}
+		output.Write(buffer[:n])
+		workingOutput := repl.Replace(output.String())
+		if prompt.MatchString(workingOutput[strings.LastIndex(workingOutput, "\n")+1:]) {
+			break
+		}
+	}
+
+	return output.String(), nil
+}
+
+// readUntilPrompt reads until the specified prompt is found and returns the read data.
+func readUntilPrompt(ctx context.Context, r io.Reader, prompt *regexp.Regexp) (string, error) {
+	var output = new(bytes.Buffer)
+	rc := newContextReader(ctx, r)
+	readSize := minReadSize
+	for {
+		b := make([]byte, readSize)
+
+		n, err := rc.Read(b)
+		if err != nil {
+			return "", fmt.Errorf("error reading output from device %w: %v", errRead, err)
+		}
+		if readSize == n && readSize < maxReadSize {
+			readSize *= 2
+		}
+		output.Write(b[:n])
+		tempSlice := bytes.ReplaceAll(output.Bytes(), []byte("\r\n"), []byte("\n"))
+		tempSlice = bytes.ReplaceAll(tempSlice, []byte("\r"), []byte("\n"))
+		if prompt.Match(tempSlice[bytes.LastIndex(tempSlice, []byte("\n"))+1:]) {
+			break
+		}
+	}
+
+	return output.String(), nil
 }