As discussed in #16, this PR tries to optimize the function reading from the ssh stdout until a prompt is detected.

Compared to what I posted in #16 I was able to further reduce the allocations (and overall performance)  by doubling the number of bytes read if the previous Read filled the buffer (bytes slice), meaning that there is probably more bytes to read. 
The doubling is capped at 10MB.

```
goos: linux
goarch: amd64
pkg: github.com/networklore/netrasp/pkg/netrasp
BenchmarkReaderOriginal100Lines            89144             13693 ns/op           51681 B/op         20 allocs/op
BenchmarkReaderStringsBuilder100Lines      77240             14421 ns/op           37649 B/op         16 allocs/op
BenchmarkReaderBytesBuffer100Lines        118660             10312 ns/op           42993 B/op         16 allocs/op

BenchmarkReaderOriginal1KLines              2932            411502 ns/op         1412229 B/op         85 allocs/op
BenchmarkReaderStringsBuilder1KLines        6068            200693 ns/op          496834 B/op         36 allocs/op
BenchmarkReaderBytesBuffer1KLines           8080            150125 ns/op          603238 B/op         34 allocs/op

BenchmarkReaderOriginal10KLines               40          29014998 ns/op        82506028 B/op        761 allocs/op
BenchmarkReaderStringsBuilder10KLines        476           2545300 ns/op         5192044 B/op         69 allocs/op
BenchmarkReaderBytesBuffer10KLines           732           1674131 ns/op         5806319 B/op         63 allocs/op

BenchmarkReaderOriginal50KLines                2         717873436 ns/op        1966611204 B/op     3955 allocs/op
BenchmarkReaderStringsBuilder50KLines         64          18521338 ns/op        31493692 B/op        102 allocs/op
BenchmarkReaderBytesBuffer50KLines            94          12372473 ns/op        37513008 B/op         94 allocs/op

BenchmarkReaderOriginal100KLines               1        3018512224 ns/op        7817554600 B/op     8220 allocs/op
BenchmarkReaderStringsBuilder100KLines        34          34133189 ns/op        62915612 B/op        112 allocs/op
BenchmarkReaderBytesBuffer100KLines           54          22204790 ns/op        68265700 B/op        102 allocs/op
PASS
ok      github.com/networklore/netrasp/pkg/netrasp      21.662s
```

I left the original function, the one using `strings.Builder` and the related benchmarks for reference. Let me know if you want me to clean them up.