@ogenstad 2021-04-20
Thanks for the PR, looks good! Yes, please go a head and clean up the reference functions and benchmarks.